eilmer: added additional preconditioners for the Newton-Krylov accelerator.
Added block-variant Jacobi and Symmetric Gauss-Seidel preconditioners.
Modified the Symmetric Gauss-Seidel relaxation preconditioner to operate on the non-transposed linear system.
Removed the transpose solve routines.

+++ src/eilmer/fluidblock.d
@@ -1043,7 +1043,7 @@ public:
 
     version(nk_accelerator) {
 
-    void initialize_jacobian(size_t spatial_order_of_jacobian)
+    void initialize_jacobian(int spatial_order_of_jacobian)
     {
         /*
           This method initializes the flow Jacobian matrix attached the FluidBlock object.
